refactor(st-09062): split tool executor modules
