+++ planning/features/09-solid-micro-refactors-feature-plan.md
@@ -2,8 +2,8 @@
 
 **Epic Range:** EP-09 through EP-09
 **Status:** In Progress
-**Last Updated:** 2026-06-10
-**Active Story:** ST-09062 - Modularize Core Tool Executor and Tests (Ready)
+**Last Updated:** 2026-06-11
+**Active Story:** ST-09062 - Modularize Core Tool Executor and Tests (In Progress)
 
 ---
